Fix E0308 errors in pluck, pollenpuff, poltergeist, powertrip
- pluck.rs: Remove Some() wrapper for single_event second argument
- pollenpuff.rs: Wrap target in Some() for heal call
- poltergeist.rs: Remove Option match, item field is ID not Option<ID>
- powertrip.rs: Cast boost values from i8 to i32

+++ src/data/move_callbacks/powertrip.rs
@@ -33,13 +33,13 @@ pub fn base_power_callback(battle: &mut Battle, pokemon_pos: (usize, usize), tar
 
             // Count all positive boosts (atk, def, spa, spd, spe, accuracy, evasion)
             let mut count = 0i32;
-            if pokemon_pokemon.boosts.atk > 0 { count += pokemon_pokemon.boosts.atk; }
-            if pokemon_pokemon.boosts.def > 0 { count += pokemon_pokemon.boosts.def; }
-            if pokemon_pokemon.boosts.spa > 0 { count += pokemon_pokemon.boosts.spa; }
-            if pokemon_pokemon.boosts.spd > 0 { count += pokemon_pokemon.boosts.spd; }
-            if pokemon_pokemon.boosts.spe > 0 { count += pokemon_pokemon.boosts.spe; }
-            if pokemon_pokemon.boosts.accuracy > 0 { count += pokemon_pokemon.boosts.accuracy; }
-            if pokemon_pokemon.boosts.evasion > 0 { count += pokemon_pokemon.boosts.evasion; }
+            if pokemon_pokemon.boosts.atk > 0 { count += pokemon_pokemon.boosts.atk as i32; }
+            if pokemon_pokemon.boosts.def > 0 { count += pokemon_pokemon.boosts.def as i32; }
+            if pokemon_pokemon.boosts.spa > 0 { count += pokemon_pokemon.boosts.spa as i32; }
+            if pokemon_pokemon.boosts.spd > 0 { count += pokemon_pokemon.boosts.spd as i32; }
+            if pokemon_pokemon.boosts.spe > 0 { count += pokemon_pokemon.boosts.spe as i32; }
+            if pokemon_pokemon.boosts.accuracy > 0 { count += pokemon_pokemon.boosts.accuracy as i32; }
+            if pokemon_pokemon.boosts.evasion > 0 { count += pokemon_pokemon.boosts.evasion as i32; }
             count
         };
